+++ compatibility/src/test/java/io/cucumber/compatibility/CompatibilityTest.java
@@ -5,8 +5,6 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.cucumber.core.options.RuntimeOptionsBuilder;
-import io.cucumber.core.plugin.HtmlFormatter;
-import io.cucumber.core.plugin.JsonFormatter;
 import io.cucumber.core.plugin.MessageFormatter;
 import io.cucumber.core.runtime.Runtime;
 import org.hamcrest.Matcher;
@@ -19,12 +17,14 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Comparator;
+import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
 import static java.nio.file.Files.newOutputStream;
+import static java.util.Collections.emptyList;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.collection.IsIterableContainingInRelativeOrder.containsInRelativeOrder;
@@ -38,8 +38,6 @@ public class CompatibilityTest {
         Path parentDir = Files.createDirectories(Paths.get("target", "messages",
             testCase.getId()));
         Path outputNdjson = parentDir.resolve("out.ndjson");
-        Path outputHtml = parentDir.resolve("out.html");
-        Path outputJson = parentDir.resolve("out.json");
 
         try {
             Runtime.builder()
@@ -48,13 +46,26 @@ public class CompatibilityTest {
                             .addFeature(testCase.getFeature())
                             .build())
                     .withAdditionalPlugins(
-                        new MessageFormatter(newOutputStream(outputNdjson)),
-                        new HtmlFormatter(newOutputStream(outputHtml)),
-                        new JsonFormatter(newOutputStream(outputJson)))
+                        new MessageFormatter(newOutputStream(outputNdjson)))
                     .build()
                     .run();
         } catch (Exception ignored) {
 
+        }
+
+        // exception: Cucumber JVM does not support named hooks
+        if ("hooks-named".equals(testCase.getId())) {
+            return;
+        }
+
+        // exception: Cucumber JVM does not support markdown features
+        if ("markdown".equals(testCase.getId())) {
+            return;
+        }
+
+        // exception: Cucumber JVM does not support retrying features
+        if ("retry".equals(testCase.getId())) {
+            return;
         }
 
         List<JsonNode> expected = readAllMessages(testCase.getExpectedFile());
@@ -63,10 +74,21 @@ public class CompatibilityTest {
         Map<String, List<JsonNode>> expectedEnvelopes = openEnvelopes(expected);
         Map<String, List<JsonNode>> actualEnvelopes = openEnvelopes(actual);
 
-        // exception: Java step definitions are not in a predictable order
-        // because Class#getMethods() does not return a predictable order.
-        sortStepDefinitions(expectedEnvelopes);
-        sortStepDefinitions(actualEnvelopes);
+        // exception: Java step definitions and hooks are not in a predictable
+        // order because Class#getMethods() does not return a predictable order.
+        sortStepDefinitionsAndHooks(expectedEnvelopes);
+        sortStepDefinitionsAndHooks(actualEnvelopes);
+
+        // exception: Cucumber JVM needs a hook to access the scenario, remove
+        // this hook from the actual test case.
+        if ("attachments".equals(testCase.getId()) || "examples-tables-attachment".equals(testCase.getId())) {
+            actualEnvelopes.getOrDefault("testCase", emptyList())
+                    .forEach(jsonNode -> {
+                        Iterator<JsonNode> testSteps = jsonNode.get("testSteps").iterator();
+                        testSteps.next();
+                        testSteps.remove();
+                    });
+        }
 
         // exception: Cucumber JVM can't execute when there are
         // unknown-parameter-types
@@ -81,7 +103,7 @@ public class CompatibilityTest {
         expectedEnvelopes.forEach((messageType, expectedMessages) -> assertThat(
             actualEnvelopes,
             hasEntry(is(messageType),
-                containsInRelativeOrder(aComparableMessage(expectedMessages)))));
+                containsInRelativeOrder(aComparableMessage(messageType, expectedMessages)))));
     }
 
     private static List<JsonNode> readAllMessages(Path output) throws IOException {
@@ -113,18 +135,30 @@ public class CompatibilityTest {
         return map;
     }
 
-    private void sortStepDefinitions(Map<String, List<JsonNode>> envelopes) {
+    private void sortStepDefinitionsAndHooks(Map<String, List<JsonNode>> envelopes) {
         Comparator<JsonNode> stepDefinitionPatternComparator = Comparator
                 .comparing(a -> a.get("pattern").get("source").asText());
         List<JsonNode> actualStepDefinitions = envelopes.get("stepDefinition");
         if (actualStepDefinitions != null) {
             actualStepDefinitions.sort(stepDefinitionPatternComparator);
         }
+        Comparator<JsonNode> hookTypeComparator = Comparator.comparing(a -> a.get("type").asText());
+        Comparator<JsonNode> hookTagExpressionComparator = Comparator.comparing(a -> {
+            JsonNode tagExpression = a.get("tagExpression");
+            if (tagExpression != null) {
+                return tagExpression.asText();
+            }
+            return "";
+        });
+        List<JsonNode> actualHooks = envelopes.get("hook");
+        if (actualHooks != null) {
+            actualHooks.sort(hookTypeComparator.thenComparing(hookTagExpressionComparator));
+        }
     }
 
-    private static List<Matcher<? super JsonNode>> aComparableMessage(List<JsonNode> messages) {
+    private static List<Matcher<? super JsonNode>> aComparableMessage(String messageType, List<JsonNode> messages) {
         return messages.stream()
-                .map(AComparableMessage::new)
+                .map(jsonNode -> new AComparableMessage(messageType, jsonNode))
                 .collect(Collectors.toList());
     }