+++ cucumber-java8/src/main/java/io/cucumber/java8/Java8DataTableTypeDefinition.java
@@ -1,0 +1,11 @@
+package io.cucumber.java8;
+
+import io.cucumber.core.backend.DataTableTypeDefinition;
+
+abstract class Java8DataTableTypeDefinition extends AbstractDatatableElementTransformerDefinition
+        implements DataTableTypeDefinition {
+
+    Java8DataTableTypeDefinition(Object body, StackTraceElement location, String[] emptyPatterns) {
+        super(body, location, emptyPatterns);
+    }
+}