+++ cucumber-java8/src/main/java/io/cucumber/java8/Java8DefaultParameterTransformerDefinition.java
@@ -3,16 +3,16 @@ package io.cucumber.java8;
 import io.cucumber.core.backend.DefaultParameterTransformerDefinition;
 import io.cucumber.cucumberexpressions.ParameterByTypeTransformer;
 
-class Java8DefaultParameterTypeDefinition extends AbstractGlueDefinition
+class Java8DefaultParameterTransformerDefinition extends AbstractGlueDefinition
         implements DefaultParameterTransformerDefinition {
 
-    Java8DefaultParameterTypeDefinition(DefaultParameterTransformerBody body) {
+    Java8DefaultParameterTransformerDefinition(DefaultParameterTransformerBody body) {
         super(body, new Exception().getStackTrace()[3]);
     }
 
     @Override
     public ParameterByTypeTransformer parameterByTypeTransformer() {
-        return (fromValue, toValue) -> invokeMethod(fromValue, toValue);
+        return this::invokeMethod;
     }
 
 }